+++ tests/.coverage
@@ -1,0 +1,1 @@
+!coverage.py: This is a private format, don't read it directly!{"lines":{"/home/arthur/Documentos/Computa\u00e7\u00e3o/7\u00ba Per\u00edodo/ES 2/trabalho/CheckErS2/src/mvc/Controller/rules.py":[2,3,4,5,6,7,9,522,11,142,403,24,25,26,28,29,286,31,419,390,302,436,184,325,70,83,469,86,87,89,91,229,100,101,486,103,104,105,106,363,108,109,110,111,112,113,114,116,374,505,102,511],"/home/arthur/Documentos/Computa\u00e7\u00e3o/7\u00ba Per\u00edodo/ES 2/trabalho/CheckErS2/src/mvc/Controller/__init__.py":[1],"/home/arthur/Documentos/Computa\u00e7\u00e3o/7\u00ba Per\u00edodo/ES 2/trabalho/CheckErS2/src/mvc/__init__.py":[1],"/home/arthur/Documentos/Computa\u00e7\u00e3o/7\u00ba Per\u00edodo/ES 2/trabalho/CheckErS2/src/mvc/Model/bot.py":[1,2,3,4,6,8,36,27,130,14],"/home/arthur/Documentos/Computa\u00e7\u00e3o/7\u00ba Per\u00edodo/ES 2/trabalho/CheckErS2/src/mvc/Model/movement.py":[64,1,130,67,132,69,128,65,10,11,140,77,142,80,24,82,78,86,88,25,26,27,29,150,96,98,37,38,39,129,41,42,43,46,111,112,113,115,40,56],"/home/arthur/Documentos/Computa\u00e7\u00e3o/7\u00ba Per\u00edodo/ES 2/trabalho/CheckErS2/src/mvc/Model/player.py":[32,1,3,40,42,13,14,50,52,53,54,55,56,57,26,27,28,29,30],"/home/arthur/Documentos/Computa\u00e7\u00e3o/7\u00ba Per\u00edodo/ES 2/trabalho/CheckErS2/src/mvc/Model/board.py":[1,59,68,69,71,8,9,79,43,44,45,47,80,17,19,56,57,27,29,31],"/home/arthur/Documentos/Computa\u00e7\u00e3o/7\u00ba Per\u00edodo/ES 2/trabalho/CheckErS2/src/mvc/Model/piece.py":[1,3,33,43,12,45,13,35,53,25,26,27,29],"/home/arthur/Documentos/Computa\u00e7\u00e3o/7\u00ba Per\u00edodo/ES 2/trabalho/CheckErS2/src/mvc/Model/__init__.py":[1]}}
